requirements.dev.txt
# SPDX-FileCopyrightText: 2023 Samsung Electronics Co., Ltd
#
# SPDX-License-Identifier: BSD-3-Clause

mypy>=1.1.1,<1.2
lxml-stubs
black
pytest>=7.2.2,<8.0.0
